requirements.txt
torch>=2.2.0
PyYAML~=6.0.3
pydantic~=2.12.5
numpy~=2.2.6
matplotlib~=3.10.8
pillow~=12.1.0
tqdm~=4.67.2

